move FD usage and file handle usage to different files
