+++ libbeat/monitoring/report/log/log.go
@@ -52,9 +52,9 @@ var gauges = map[string]bool{
 	"beat.cpu.total.value":           true,
 	"beat.cpu.total.ticks":           true,
 	"beat.cpu.total.time":            true,
-	"beat.fd.open":                   true,
-	"beat.fd.limit.hard":             true,
-	"beat.fd.limit.soft":             true,
+	"beat.handles.open":              true,
+	"beat.handles.limit.hard":        true,
+	"beat.handles.limit.soft":        true,
 	"system.load.1":                  true,
 	"system.load.5":                  true,
 	"system.load.15":                 true,
